pyproject.toml
[project]
name = "infinitas-timestamper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.182.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "injector>=0.22.0",
    "obs-websocket-py>=1.0",
    "obsws-python>=1.8.0",
    "pydantic>=2.11.9",
    "pyperclip>=1.10.0",
    "pyside6[tools]>=6.9.2",
    "watchdog>=6.0.0",
]
